refactor!: reorganize source code
BREAKING CHANGE: This does not actually break any interfaces, but would be breaking if you are importing from deeply nested paths which is not supported.
