--- src/github-handler/index.ts
@@ -1,20 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-export * from './fork-handler';
-export {branch} from './branch-handler';
-export {commitAndPush} from './commit-and-push-handler';
-export * from './pull-request-handler';
-export * from './comment-handler';
-export * from './issue-handler';
