+++ src/utils/hunk-utils.ts
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//   https://www.apache.org/licenses/LICENSE-2.0
+//    http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,9 +12,73 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {Hunk} from '../../../types';
-import {adjustHunkUp, adjustHunkDown} from '../../hunk-utils';
+import {Hunk, FileDiffContent} from '../types';
+import {getSuggestedHunks} from './diff-utils';
+import {logger} from '../logger';
 
+/**
+ * Shift a Hunk up one line so it starts one line earlier.
+ * @param {Hunk} hunk
+ * @returns {Hunk | null} the adjusted Hunk or null if there is no preceeding line.
+ */
+export function adjustHunkUp(hunk: Hunk): Hunk | null {
+  if (!hunk.previousLine) {
+    return null;
+  }
+  return {
+    oldStart: hunk.oldStart - 1,
+    oldEnd: hunk.oldEnd,
+    newStart: hunk.newStart - 1,
+    newEnd: hunk.newEnd,
+    newContent: [hunk.previousLine, ...hunk.newContent],
+  };
+}
+
+/**
+ * Shift a Hunk up one line so it ends one line later.
+ * @param {Hunk} hunk
+ * @returns {Hunk | null} the adjusted Hunk or null if there is no following line.
+ */
+export function adjustHunkDown(hunk: Hunk): Hunk | null {
+  if (!hunk.nextLine) {
+    return null;
+  }
+  return {
+    oldStart: hunk.oldStart,
+    oldEnd: hunk.oldEnd + 1,
+    newStart: hunk.newStart,
+    newEnd: hunk.newEnd + 1,
+    newContent: hunk.newContent.concat(hunk.nextLine),
+  };
+}
+
+/**
+ * Given a map where the key is the file name and the value is the
+ * old content and new content of the file
+ * compute the hunk for each file whose old and new contents differ.
+ * Do not compute the hunk if the old content is the same as the new content.
+ * The hunk list is sorted and each interval is disjoint.
+ * @param {Map<string, FileDiffContent>} diffContents a map of the original file contents and the new file contents
+ * @returns the hunks for each file whose old and new contents differ
+ */
+export function getRawSuggestionHunks(
+  diffContents: Map<string, FileDiffContent>
+): Map<string, Hunk[]> {
+  const fileHunks: Map<string, Hunk[]> = new Map();
+  diffContents.forEach((fileDiffContent: FileDiffContent, fileName: string) => {
+    // if identical don't calculate the hunk and continue in the loop
+    if (fileDiffContent.oldContent === fileDiffContent.newContent) {
+      return;
+    }
+    const hunks = getSuggestedHunks(
+      fileDiffContent.oldContent,
+      fileDiffContent.newContent
+    );
+    fileHunks.set(fileName, hunks);
+  });
+  logger.info('Parsed ranges of old and new patch');
+  return fileHunks;
+}
 interface PartitionedHunks {
   validHunks: Map<string, Hunk[]>;
   invalidHunks: Map<string, Hunk[]>;
